#795 Create new router for all deadline notifications, more abstraction
