+++ .github/workflows/task-deadline-notifications.yml
@@ -10,5 +10,5 @@ jobs:
     steps:
       - name: Send notifications
         run: |
-          curl -X GET https://api.finishlinebyner.com/tasks/sendTaskDeadlineSlackNotifications \
+          curl -X GET https://api.finishlinebyner.com/deadline-notifications/sendTaskDeadlineSlackNotifications \
           -H 'Authorization: ${{ secrets.NOTIFICATION_ENDPOINT_SECRET }}'
