+++ src/backend/index.ts
@@ -12,6 +12,7 @@ import descriptionBulletsRouter from './src/routes/description-bullets.routes';
 import tasksRouter from './src/routes/tasks.routes';
 import reimbursementRequestsRouter from './src/routes/reimbursement-requests.routes';
 import designReviewRouter from './src/routes/design-review.routes';
+import deadlineNotificationsRouter from './src/routes/deadline-notifications.routes';
 
 const app = express();
 const port = process.env.PORT || 3001;
@@ -54,6 +55,7 @@ app.use('/description-bullets', descriptionBulletsRouter);
 app.use('/tasks', tasksRouter);
 app.use('/reimbursement-requests', reimbursementRequestsRouter);
 app.use('/design-reviews', designReviewRouter);
+app.use('/deadline-notifications', deadlineNotificationsRouter);
 app.use('/', (_req, res) => {
   res.json('Welcome to FinishLine');
 });
